+++ cppcache/include/geode/Serializer.hpp
@@ -117,7 +117,7 @@ inline void writeObject(apache::geode::client::DataOutput& output,
 
 inline void readObject(apache::geode::client::DataInput& input,
                        uint16_t& value) {
-  value = static_cast<uint16_t>(input.readInt16());
+  value = input.readInt16();
 }
 
 inline void writeObject(apache::geode::client::DataOutput& output,
